Moved <SearchFormSync /> out of the pack file.
