+++ app/javascript/Search/index.js
@@ -1,2 +1,3 @@
 export * from './Search';
 export * from './SearchForm';
+export * from './SearchFormSync';
